# docs/projects/api/abodemine-openapi-3.0.0.yaml
openapi: 3.0.0
info:
  title: AbodeMine API
  version: 3.0.0
  description: API for retrieving real estate data with various filter options and address search capabilities.
servers:
  - url: https://api.abodemine.com/api/v3
    description: Production API
security:
  - bearerAuth: []
paths:
  /listings:
    post:
      summary: Get listings based on filter criteria
      description: >
        Retrieve real estate listings based on various filtering options.


        ## Authentication

        All requests require API authentication using a Bearer token with the API key in the header.


        ## Filter Parameters

        This endpoint supports extensive filtering capabilities to narrow down your search results. You can filter by:


        - **Property characteristics** (bedrooms, bathrooms, square footage)

        - **Price range** (minimum and maximum listing price)

        - **Listing details** (status, year built)

        - **Location** (geographic coordinates, ZIP codes)

        - **Specific listings** (MLS numbers with OUID)


        All parameters are optional but providing more specific filters will yield more targeted results.


        ## Search Methods

        ### Polygon Search

        Define a polygon by providing an array of latitude/longitude points within the geoPolygon object. The search will return properties located within this polygon. For accurate results, ensure the polygon is closed by making the first and last points identical.


        **Note:** The polygon search has a maximum area limit of 30.0k miles.


        **Example usage polygon:**

        ```json

        {
          "geoFilter": {
            "geoPolygon": {
              "points": [
                {
                  "lat": 42.151677,
                  "lon": -88.279575
                },
                {
                  "lat": 42.150677,
                  "lon": -88.281268
                },
                {
                  "lat": 42.149691,
                  "lon": -88.278443
                },
                {
                  "lat": 42.151677,
                  "lon": -88.279575
                }
              ]
            }
          },
          "minBeds": 3,
          "maxPrice": 750000
        }

        ```


        ### Radius Search

        Define a central point and radius to search within a circular area. The search will return properties located within the specified distance from the central point.


        **Note:** The radius is measured in miles, with a maximum limit of 100.0 miles.


        **Example usage radius by lat/lon:**

        ```json

        {
          "geoFilter": {
            "geoDistance": {
              "location": {
                "lat": 34.045009,
                "lon": -118.452457
              },
              "radius": 5
            }
          }
        }

        ```


        **Example usage radius by address:**

        ```json

        {
          "geoFilter": {
            "address": {
              "fullStreetAddress": "10656 DRAKEWOOD AVE",
              "city": "Culver City",
              "state": "CA"
            },
            "geoDistance": {
              "radius": 5
            }
          },
          "minYearBuilt": 1980,
          "statuses": ["Active"]
        }

        ```

          
        **Example usage radius by aupid:**

        ```json

        {
          "geoFilter": {
            "aupid": "400f91c0-4ca0-4295-a2c3-8ebd59560b87",
            "geoDistance": {
              "radius": 5
            }
          }
        }

        ```


        ### MLS Number Search

        Search for specific MLS listings by providing both the OUID (Originating MLS Unique Identifier) and one or more MLS numbers.


        **Note:** When searching by MLS numbers, you must provide both the ouid and at least one MLS number. You can find the appropriate OUID for each MLS using the [MLS Sources lookup table](https://docs.abodemine.com/docs/mls-sources).

        **Example usage:**

        ```json

        {
          "ouid": "M00000650",
          "mlsNumbers": [
            "2333737",
            "2330147",
            "1772422"
          ]
        }

        ```


        ### Basic Property Filter

        Filter listings based on property characteristics, price range, and listing status.

        **Example usage:**

        ```json

        {
          "minBeds": 2,
          "maxBeds": 4,
          "minPrice": 200000,
          "maxPrice": 500000,
          "zip5Codes": ["98102", "98103"],
          "statuses": ["Active", "Pending"],
          "pageLimit": 100,
          "pageNumber": 0
        }

        ```

        ### Assessor Property Subtype Filter

        Filter listings based on [assessor property subtypes](https://docs.abodemine.com/docs/assessor-property-sub-types).

        **Example usage:**

        ```json

        {
          "minBeds": 2,
          "maxBeds": 4,
          "minPrice": 200000,
          "maxPrice": 500000,
          "zip5Codes": ["98102", "98103"],
          "statuses": ["Active", "Pending"],
          "asrPropertySubType": "Single Family Residence",
          "pageLimit": 100,
          "pageNumber": 0
        }

        ```


        ## Photo URL Construction

        To access listing photos, construct the URL using this pattern:


        ```
        {photoUrlPrefix}/{photoKey}/photo_{index}.jpg
        ```


        Where:


        `photoUrlPrefix`: The base URL prefix from the listing response
        
        `photoKey`: The unique photo key identifier from the listing
        
        `index`: Photo number starting from 1 up to `photosCount`


        Example:

        ```

        https://d2sa73fkuaa019.cloudfront.net/listings/nwmls/NWM119125689/photo_1.jpg

        ```
      operationId: getListings
      security:
        - bearerAuth: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                aupid:
                  type: string
                  description: AbodeMine Data's unique parcel identifier
                  example: "7e9aa888-e1fd-4cc1-9b20-ba47269f90cf"
                minBeds:
                  type: integer
                  description: Minimum number of bedrooms
                  example: 2
                maxBeds:
                  type: integer
                  description: Maximum number of bedrooms
                  example: 4
                minBaths:
                  type: integer
                  description: Minimum number of full bathrooms
                  example: 1
                maxBaths:
                  type: integer
                  description: Maximum number of full bathrooms
                  example: 3
                minSqFt:
                  type: integer
                  description: Minimum living area square footage
                  example: 1000
                maxSqFt:
                  type: integer
                  description: Maximum living area square footage
                  example: 3000
                minPrice:
                  type: integer
                  description: Minimum listing price in dollars
                  example: 200000
                maxPrice:
                  type: integer
                  description: Maximum listing price in dollars
                  example: 500000
                minYearBuilt:
                  type: integer
                  description: Minimum year built
                  example: 1980
                maxYearBuilt:
                  type: integer
                  description: Maximum year built
                  example: 2023
                minStatusChangeDate:
                  type: string
                  format: date
                  description: Minimum status change date in ISO 8601 format
                  example: "2023-01-01"
                statuses:
                  type: array
                  description: Filter results by listing statuses (e.g., "Active", "Pending", "Sold")
                  items:
                    type: string
                  example: ["Active", "Pending"]
                zip5Codes:
                  type: array
                  description: Filter results by one or more 5-digit ZIP codes
                  items:
                    type: string
                  example: ["98102", "98103"]
                ouid:
                  type: string
                  description: Originating MLS unique identifier.
                  example: "M00000650"
                mlsNumbers:
                  type: array
                  description: Array of MLS listing numbers to retrieve
                  items:
                    type: string
                  example: ["2333737", "2330147", "1772422"]                  
                pageLimit:
                  type: integer
                  description: Maximum number of results to return per page (default 10, max 100)
                  default: 10
                  maximum: 100
                  example: 100
                pageNumber:
                  type: integer
                  description: Page number for paginated results (starts at 0)
                  default: 0
                  example: 0
                geoFilter:
                  type: object
                  description: Filter results by geographic area using either polygon coordinates or a radius search. See the Search Methods section for detailed examples.
                  oneOf:
                    - type: object
                      required:
                        - geoPolygon
                      properties:
                        geoPolygon:
                          type: object
                          description: Polygon shape definition for geographic filtering
                          properties:
                            points:
                              type: array
                              description: Array of geographic coordinates that define a polygon shape
                              items:
                                type: object
                                required:
                                  - lat
                                  - lon
                                properties:
                                  lat:
                                    type: number
                                    description: Latitude coordinate in decimal degrees
                                    example: 42.151677
                                  lon:
                                    type: number
                                    description: Longitude coordinate in decimal degrees
                                    example: -88.279575
                    - type: object
                      required:
                        - geoDistance
                      properties:
                        geoDistance:
                          type: object
                          description: Circular radius search definition for geographic filtering
                          required:
                            - radius
                          properties:
                            radius:
                              type: number
                              description: Search radius in miles (max 100.0)
                              example: 5.0
                              maximum: 100.0
                        address:
                          type: object
                          description: Address object for radius search center point
                          properties:
                            fullStreetAddress:
                              type: string
                              description: Full street address
                              example: "10656 DRAKEWOOD AVE"
                            houseNumber:
                              type: string
                              description: House number component
                              example: "10656"
                            streetPreDirection:
                              type: string
                              description: Street pre-direction component (NE, SW, etc.)
                              example: "NE"
                            streetName:
                              type: string
                              description: Street name component
                              example: "DRAKEWOOD"
                            streetSuffix:
                              type: string
                              description: Street suffix component (Ave, St, Blvd, etc.)
                              example: "Ave"
                            streetPostDirection:
                              type: string
                              description: Street post-direction component if exists
                              example: "SW"
                            unitType:
                              type: string
                              description: Unit type component if exists
                              example: "Apt"
                            unitNumber:
                              type: string
                              description: Unit number component if exists
                              example: "2B"
                            city:
                              type: string
                              description: City name
                              example: "Culver City"
                            state:
                              type: string
                              description: State code (two-letter)
                              example: "CA"
                            zip5:
                              type: string
                              description: ZIP code
                              example: "90230"
                        aupid:
                          type: string
                          description: AbodeMine Data's unique parcel identifier for radius search center point
                          example: "400f91c0-4ca0-4295-a2c3-8ebd59560b87"
                        location:
                          type: object
                          description: Latitude/longitude coordinates for radius search center point
                          properties:
                            lat:
                              type: number
                              description: Latitude coordinate of the center point
                              example: 34.052235
                            lon:
                              type: number
                              description: Longitude coordinate of the center point
                              example: -118.243683      
      responses:
        "200":
          description: >
            # Successful Response


            The response includes an array of listing objects and pagination information.


            ## Example Response


            ```json

            {
              "listings": [
                {
                  "statusChangeDate": "2024-11-22T00:00:00Z",
                  "propertyAddressFull": "1330 MEGHAN AVE",
                  "propertyAddressCity": "ALGONQUIN",
                  "propertyAddressState": "IL",
                  "propertyAddressZip": "60102",
                  "listingStatus": "Sold",
                  "mlsSoldPrice": 340000,
                  "yearBuilt": 1984,
                  "livingAreaSquareFeet": 1194,
                  "bedroomsTotal": 3,
                  "bathroomsFull": "1",
                  "bathroomsHalf": 1
                }
              ],
              "pagination": {
                "pageNumber": 1,
                "pageLimit": 10,
                "totalResults": 150
              }
            }

            ```


            ## Response Fields


            See the Listing schema below for detailed field descriptions.
          content:
            application/json:
              schema:
                type: object
                properties:
                  listings:
                    type: array
                    items:
                      $ref: "#/components/schemas/Listing"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          description: Bad Request - Invalid parameters
        "401":
          description: Unauthorized - Authentication required
        "500":
          description: Internal Server Error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Listing:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the listing
        statusChangeDate:
          type: string
          format: date-time
          description: The transactional timestamp documenting the date and time for every record change
        propertyAddressFull:
          type: string
          description: Full standardized site address line (e.g., '123 1/2 N Main St')
        propertyAddressHouseNumber:
          type: string
          description: Standardized site address house number and fraction
        propertyAddressStreetName:
          type: string
          description: Standardized site address street name
        propertyAddressStreetSuffix:
          type: string
          description: Standardized site address street name suffix
        propertyAddressCity:
          type: string
          description: Standardized site address city name
        propertyAddressState:
          type: string
          description: Standardized site address state
        propertyAddressZip:
          type: string
          description: Standardized site address ZIP Code
        propertyAddressZip4:
          type: string
          description: Standardized site address ZIP Plus 4 code
        situsCounty:
          type: string
          description: County where the property is situated
        township:
          type: string
          description: A subdivision of the county
        mlsListingAddress:
          type: string
          description: The full listing address, unparsed, including street name, number and direction
        mlsListingCity:
          type: string
          description: The city in the listing address
        mlsListingState:
          type: string
          description: The state in the listing address
        mlsListingZip:
          type: string
          description: The postal ZIP code in the listing address
        mlsListingCountyFips:
          type: string
          description: The county, parish or other regional authority where the listing address is located
        mlsNumber:
          type: string
          description: The number used to identify a listing during its contract period as sourced by the MLS
        mlsSource:
          type: string
          description: The name of the Originating record provider, most commonly the name of the MLS
        listingStatus:
          type: string
          description: >-
            The status of the listing (Active, Active Under Contract, Canceled, Closed, Coming Soon, Delete, Expired, Hold, Incomplete, Pending, Withdrawn, Sold)
          enum:
            - Active
            - Active Under Contract
            - Canceled
            - Closed
            - Coming Soon
            - Delete
            - Expired
            - Hold
            - Incomplete
            - Pending
            - Withdrawn
            - Sold
        mlsSoldDate:
          type: string
          format: date-time
          description: With for-sale listings, the date the purchase agreement was fulfilled
        mlsSoldPrice:
          type: integer
          description: The amount of money paid by the purchaser to the seller for the property under the agreement
        assessorLastSaleDate:
          type: string
          format: date-time
          description: Date which the primary owner acquired the property as provided by the assessor
        assessorLastSaleAmount:
          type: string
          description: Amount paid by primary owner as provided by the assessor
        avgMarketPricePerSqFt:
          type: integer
          description: The average market price per square foot for residential properties
        listingDate:
          type: string
          format: date-time
          description: The date the listing was placed on market
        latestListingPrice:
          type: integer
          description: The most recent listing price for the property
        previousListingPrice:
          type: integer
          description: The most recent previous ListingPrice of the listing
        specialListingConditions:
          type: string
          description: A list of options that describe the type of sale (Standard, REO, Short Sale, Probate, Auction, NOD, etc.)
        originalListingDate:
          type: string
          format: date-time
          description: The transactional timestamp when the listing was entered and made visible to members of the MLS
        originalListingPrice:
          type: integer
          description: The original price of the property on the initial agreement between the seller and the broker
        concessions:
          type: string
          description: Indicates if there are any seller concessions agreed upon between the Buyer and Seller
        mlsPropertyType:
          type: string
          description: A list of types of properties such as Residential, Lease, Income, Land, Mobile, Commercial Sale, etc.
        mlsPropertySubType:
          type: string
          description: A list of types of residential and residential lease properties, i.e. SFR, Condo, etc.
        propertyType:
          type: string
          description: Standardized value to describe property use, derived from jurisdiction-specific zoned use
        propertySubType:
          type: string
          description: General property type description (residential, commercial, other, etc.) as derived from the assessor
        ownerOccupied:
          type: boolean
          description: Whether the property is owner-occupied
        assdTotalValue:
          type: integer
          description: Assessed total value
        marketTotalValue:
          type: integer
          description: Market total value
        marketValueLand:
          type: integer
          description: Market value of the land
        marketValueImprovement:
          type: integer
          description: Market value of improvements
        taxAmt:
          type: integer
          description: Tax amount
        taxYear:
          type: integer
          description: Tax year
        marketYear:
          type: integer
          description: Market valuation year
        sumLivingAreaSqFt:
          type: integer
          description: Total living area in square feet
        sumBasementSqFt:
          type: integer
          description: Total basement area in square feet
        basementFinishedSqFt:
          type: integer
          description: Finished basement area in square feet
        sumGarageSqFt:
          type: integer
          description: Total garage area in square feet
        yearBuilt:
          type: integer
          description: Year the property was built
        bedrooms:
          type: string
          description: Number of bedrooms
        totalRooms:
          type: string
          description: Total number of rooms
        bathTotalCalc:
          type: string
          description: Calculated total bathrooms
        bathFull:
          type: string
          description: Number of full bathrooms
        bathsPartialNbr:
          type: string
          description: Number of partial bathrooms
        shipDescription:
          type: string
          description: Identifies the ownership/interest rights to the property (e.g. Fee Simple, Condominium, Cooperative)
        latitude:
          type: number
          format: double
          description: The geographic latitude of a reference point on the property, specified in degrees
        longitude:
          type: number
          format: double
          description: The geographic longitude of a reference point on the property, specified in degrees
        apnFormatted:
          type: string
          description: Primary Parcel Number and unique identifier within the county/jurisdiction
        legalDescription:
          type: string
          description: Legal description of the property
        legalSubdivision:
          type: string
          description: The name of the legal subdivision for the property, typically a neighborhood or community
        daysOnMarket:
          type: integer
          description: The number of days the listing is on market
        listingAgentFullName:
          type: string
          description: The full name of the listing agent
        listingAgentMlsId:
          type: string
          description: The local, well-known identifier for the listing agent
        listingAgentPreferredPhone:
          type: string
          description: Preferred phone number for contacting the listing agent
        listingAgentEmail:
          type: string
          description: The email address of the Listing Agent
        listingOfficeName:
          type: string
          description: The legal name of the brokerage representing the seller
        listingOfficeMlsId:
          type: string
          description: The local, well-known identifier for the listing office
        listingOfficePhone:
          type: string
          description: The primary phone number for the listing agent's office
        listingOfficeEmail:
          type: string
          description: The email address of the Listing Office
        buyerAgentFullName:
          type: string
          description: The full name of the buyer's agent
        buyerAgentMlsId:
          type: string
          description: The local, well-known identifier for the buyer's agent
        buyerAgentPreferredPhone:
          type: string
          description: Preferred contact phone number for the buyer agent involved in the transaction
        buyerAgentEmail:
          type: string
          description: The email address of the Buyer's Agent
        buyerOfficeName:
          type: string
          description: The legal name of the brokerage representing the buyer
        buyerOfficeMlsId:
          type: string
          description: The local, well-known identifier for the buyer's office
        buyerOfficePhone:
          type: string
          description: The primary phone number for the buyer agent's office
        buyerOfficeEmail:
          type: string
          description: The email address of the Buyer's Office
        publicListingRemarks:
          type: string
          description: Public remarks about the listing that describe the property features and amenities
        taxYearAssessed:
          type: integer
          description: The tax year for which the property was last assessed
        taxAmount:
          type: integer
          description: The amount of property taxes assessed
        ownerName:
          type: string
          description: Name of the property owner
        ownerVesting:
          type: string
          description: How the owner holds title to the property
        lotDimensions:
          type: string
          description: Dimensions of the property lot
        lotFeatureList:
          type: string
          description: List of features of the property lot
        livingAreaSquareFeet:
          type: integer
          description: Total living area of the property in square feet
        livingAreaSource:
          type: string
          description: Source of the living area measurement
        roomsTotal:
          type: integer
          description: Total number of rooms in the property
        bedroomsTotal:
          type: integer
          description: Total number of bedrooms in the property
        bathroomsFull:
          type: string
          description: Number of full bathrooms in the property
        bathroomsHalf:
          type: integer
          description: Number of half bathrooms in the property
        basementFeatures:
          type: string
          description: Description of the basement features
        repairs:
          type: boolean
          description: Indicates if repairs are needed for the property
        garageSpaces:
          type: string
          description: Number of garage spaces available
        parkingTotal:
          type: integer
          description: Total number of parking spaces available
        heatingFeatures:
          type: string
          description: Features of the heating system
        cooling:
          type: string
          description: Type of cooling system in the property
        fireplace:
          type: string
          description: Type of fireplace in the property
        fireplaceNumber:
          type: integer
          description: Number of fireplaces in the property
        architecturalStyleFeatures:
          type: string
          description: Architectural style of the property
        waterSource:
          type: string
          description: Source of water for the property
        sewer:
          type: string
          description: Type of sewer system for the property
        otherEquipmentIncluded:
          type: string
          description: Other equipment included with the property
        laundryFeatures:
          type: string
          description: Features of the laundry facilities
        appliances:
          type: string
          description: Appliances included with the property
        interiorFeatures:
          type: string
          description: Features of the interior of the property
        exteriorFeatures:
          type: string
          description: Features of the exterior of the property
        association1FeeFrequency:
          type: string
          description: Frequency of HOA or condo association fees
        associationFeeIncludes:
          type: string
          description: What is included in the association fees
        schoolElementary:
          type: string
          description: Name of the elementary school for the property
        schoolElementaryDistrict:
          type: string
          description: School district for the elementary school
        schoolMiddle:
          type: string
          description: Name of the middle school for the property
        schoolMiddleDistrict:
          type: string
          description: School district for the middle school
        schoolHigh:
          type: string
          description: Name of the high school for the property
        schoolHighDistrict:
          type: string
          description: School district for the high school
        listingMarketingUrl:
          type: string
          description: URL for marketing materials related to the listing
        photosCount:
          type: integer
          description: Number of photos available for the listing
        photoKey:
          type: string
          description: Key used to access photos of the listing
        photoUrlPrefix:
          type: string
          description: URL prefix for accessing listing photos
        aupid:
          type: string
          nullable: true
          description: AbodeMine Data's unique parcel identifier
        relationToSubject:
          type: string
          description: Relation of the listing to the subject property
    Pagination:
      type: object
      properties:
        pageNumber:
          type: integer
          description: Current page number
        pageLimit:
          type: integer
          description: Number of results per page
        totalResults:
          type: integer
          description: Total number of results matching the query
...